Opens a connection to an sqlite database.

The connection can be queried and closed using the functions in the DBI package.

ps_connect_sqlite(file = "database", dir = ".", new = NA,
  foreign_keys = TRUE, ask = getOption("poissqlite.ask", TRUE))

Arguments

file
A string of the name of the file (without the file extension .sqlite).

dir
A string of the directory.

new
A flag indicating whether to enforce connection to an existing database (FALSE) or connection to a new database (TRUE).

foreign_keys
A flag indicating whether to switch foreign keys on.

ask
A flag indicating whether to ask before creating a new directory.